Refactor: split lib.rs into logical modules
Split the 1319-line lib.rs into focused modules for better organization
and maintainability:

- **lib.rs** (186 lines, -86%): DagShell core, module declarations
- **shell_ui.rs** (217 lines): CLI parsing, formatting, help text,
  ChannelSink for rustyline integration
- **output.rs** (231 lines): OutputSink trait, 256-color support,
  line-buffering OutputSinkWriter
- **ailetos_rt_glue.rs** (162 lines): Event bridging, Ctrl+C handler,
  notification watcher, executor startup
- **commands.rs** (673 lines): All 21 cmd_* implementations
- **main.rs** (69 lines, -50%): Simplified using shell_ui utilities

Module names clarify purpose:
- "ailetos_rt_glue" = glue between sync CLI and async ailetos runtime
- "shell_ui" = user-facing strings, parsing, formatting

Benefits:
- Clear separation of concerns
- Easier code navigation and review
- Explicit module dependencies
- No behavioral changes

All tests pass.
